+++ telepot/aio/__init__.py
@@ -105,6 +105,12 @@ class Bot(_BotBase):
         p = _strip(locals())
         return await self._api_request('forwardMessage', _rectify(p))
 
+    async def copyMessage(self, chat_id, from_chat_id, message_id,
+                             disable_notification=None):
+        """ See: https://core.telegram.org/bots/api#copyMessage """
+        p = _strip(locals())
+        return await self._api_request('copyMessage', _rectify(p))
+
     async def sendPhoto(self, chat_id, photo,
                         caption=None,
                         parse_mode=None,
@@ -329,11 +335,11 @@ class Bot(_BotBase):
         p = _strip(locals())
         return await self._api_request('getFile', _rectify(p))
 
-    async def kickChatMember(self, chat_id, user_id,
+    async def banChatMember(self, chat_id, user_id,
                              until_date=None):
-        """ See: https://core.telegram.org/bots/api#kickchatmember """
+        """ See: https://core.telegram.org/bots/api#banChatMember """
         p = _strip(locals())
-        return await self._api_request('kickChatMember', _rectify(p))
+        return await self._api_request('banChatMember', _rectify(p))
 
     async def unbanChatMember(self, chat_id, user_id):
         """ See: https://core.telegram.org/bots/api#unbanchatmember """